Move to_snake_keys method in Plissken::Methods
