+++ lib/plissken.rb
@@ -1,5 +1,5 @@
-require File.dirname(__FILE__) + '/plissken/ext/hash/to_snake_keys'
+require "plissken/methods"
+require "plissken/ext/hash/to_snake_keys"
 
 module Plissken
-
 end
